+++ codehunt/front-end/js/profile.js
@@ -136,29 +136,44 @@ function submitFormUser(){
 function removeItem(){
   event.preventDefault();
   var itemToRemove = $(this).parent().parent().parent();
-  $.ajax({
-    url:'http://localhost:3000/'+$(this).data().id,
-    type:'delete',
-    beforeSend: setRequestHeader
-  }).done(function() {
-    itemToRemove.remove();
-  });
+  var urldata = $(this).data().id;
+
+  swal({   title: "Are you sure you want to delete this post?",   showCancelButton: true,   confirmButtonColor: "#DD6B55",   confirmButtonText: "Delete",   cancelButtonText: "Cancel",   closeOnConfirm: true,   closeOnCancel: true }, function(isConfirm){   if (isConfirm) {     
+    
+    $.ajax({
+      url:'http://localhost:3000/'+urldata,
+      type:'delete',
+      beforeSend: setRequestHeader
+    }).done(function() {
+      itemToRemove.remove();
+    });  
+
+  }})   
 }
 
 // REMOVE item - user
 function removeUser(){
+
   event.preventDefault();
   var itemToRemove = $(this).parent();
-  $.ajax({
-    url:'http://localhost:3000/profile/'+$(this).data().id,
-    type:'delete',
-    beforeSend: setRequestHeader
-  }).done(function() {
-    itemToRemove.remove();
-    removeToken();
-    loggedOutState();
-    window.location.href = 'index.html';
-  });
+  var urldata = $(this).data().id;
+
+  swal({   title: "Are you sure you want to delete your account?",   showCancelButton: true,   confirmButtonColor: "#DD6B55",   confirmButtonText: "Delete",   cancelButtonText: "Cancel",   closeOnConfirm: true,   closeOnCancel: true }, function(isConfirm){   if (isConfirm) {     
+    
+    $.ajax({
+      url:'http://localhost:3000/profile/'+ urldata,
+      type:'delete',
+      beforeSend: setRequestHeader
+    }).done(function() {
+      itemToRemove.remove();
+      removeToken();
+      loggedOutState();
+      window.location.href = 'index.html';
+    }); 
+
+  }})   
+
+
 }
 
 // EDIT user
